Added the swagger config in Diff file
