+++ app/controllers/app_controllers.py
@@ -1,6 +1,5 @@
 import os
 from fastapi import APIRouter, HTTPException, Depends
-from fastapi.security import OAuth2PasswordBearer
 from pydantic import BaseModel
 from typing import Union
 from app.db.mongodata import db
@@ -23,15 +22,6 @@ class Item(BaseModel):
 
 router = APIRouter()
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/token")
-
-@router.get("/protected", tags=["Protected"])
-async def protected_endpoint(token: str = Depends(oauth2_scheme)):
-    """
-    A protected endpoint that requires authentication.
-    """
-    return {"message": "You are authenticated!", "token": token}
-
 
 @router.get("/healthcheck", tags=["Health"])
 async def healthcheck():
@@ -41,27 +31,27 @@ async def healthcheck():
     return {"status": "ok"}
 
 
-@router.get("/items/{item_id}", tags=["Items"])
-async def get_item(item_id: int, current_user: dict = Depends(get_current_user)):
-    """
-    Retrieve an item by its ID. Requires authentication.
-    """
-    if item_id <= 0:
-        raise HTTPException(status_code=400, detail="Invalid item ID")
-    return {"item_id": item_id, "name": f"Item {item_id}", "user": current_user}
+# @router.get("/items/{item_id}", tags=["API"])
+# async def get_item(item_id: int, current_user: dict = Depends(get_current_user)):
+#     """
+#     Retrieve an item by its ID. Requires authentication.
+#     """
+#     if item_id <= 0:
+#         raise HTTPException(status_code=400, detail="Invalid item ID")
+#     return {"item_id": item_id, "name": f"Item {item_id}", "user": current_user}
 
 
-@router.post("/items", tags=["Items"])
-async def create_item(item: dict, current_user: dict = Depends(get_current_user)):
-    """
-    Create a new item. Requires authentication.
-    """
-    if "name" not in item or not item["name"]:
-        raise HTTPException(status_code=400, detail="Item name is required")
-    return {"message": "Item created successfully", "item": item, "user": current_user}
+# @router.post("/items", tags=["API"])
+# async def create_item(item: dict, current_user: dict = Depends(get_current_user)):
+#     """
+#     Create a new item. Requires authentication.
+#     """
+#     if "name" not in item or not item["name"]:
+#         raise HTTPException(status_code=400, detail="Item name is required")
+#     return {"message": "Item created successfully", "item": item, "user": current_user}
 
 
-@router.post("/dbitems/", tags=["Items"])
+@router.post("/items/", tags=["API"])
 async def create_db_item(item: dict, current_user: dict = Depends(get_current_user)):
     """
     Create a new item in the database. Requires authentication.
@@ -70,7 +60,7 @@ async def create_db_item(item: dict, current_user: dict = Depends(get_current_us
     return {"message": "Item created", "id": str(result.inserted_id), "user": current_user}
 
 
-@router.get("/dbitems/{item_id}", tags=["Items"])
+@router.get("/items/{item_id}", tags=["API"])
 async def get_db_item(item_id: str, current_user: dict = Depends(get_current_user)):
     """
     Retrieve an item from the database by its ID. Requires authentication.
