+++ app/utils/swagger.py
@@ -1,0 +1,35 @@
+from fastapi import FastAPI
+from fastapi.openapi.utils import get_openapi
+from fastapi.security import OAuth2PasswordBearer
+
+# Define OAuth2 scheme here
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/token")
+
+# Custom OpenAPI schema function
+def custom_openapi(app: FastAPI):
+    if app.openapi_schema:
+        return app.openapi_schema
+
+    openapi_schema = get_openapi(
+        title="FastAPI Application",
+        version="1.0.0",
+        description="This is a sample FastAPI application with authentication.",
+        routes=app.routes,
+    )
+    openapi_schema["components"]["securitySchemes"] = {
+        "OAuth2PasswordBearer": {
+            "type": "oauth2",
+            "flows": {
+                "password": {
+                    "tokenUrl": "/auth/token",
+                    "scopes": {}
+                }
+            }
+        }
+    }
+    for path in openapi_schema["paths"].values():
+        for method in path.values():
+            method["security"] = [{"OAuth2PasswordBearer": []}]
+    
+    app.openapi_schema = openapi_schema
+    return app.openapi_schema
